+++ lib/src/commands.rs
@@ -248,7 +248,7 @@ impl<'r, 's: 'r> Command<'r, 's> {
     fn new(name: &'s str, options: Option<&'r OptionSet<'r, 's>>, sub_commands: CommandSet<'r, 's>) -> Self {
         debug_assert!(!name.is_empty(), "Command name cannot be an empty string!");
         debug_assert!(!name.contains('\u{FFFD}'), "Command name cannot contain ‘\\u{FFFD}’!");
-        let opts_actual = options.unwrap_or(&gong_option_set_fixed!());
+        let opts_actual = options.unwrap_or(&gong_option_set!());
         Self { name, options: opts_actual, sub_commands }
     }
 }
@@ -350,7 +350,7 @@ mod tests {
     #[test]
     #[cfg_attr(debug_assertions, should_panic)]
     fn create_cmd_no_name() {
-        let _cmd = Command::new("", None, gong_command_set_fixed!()); // Should panic here in debug mode!
+        let _cmd = Command::new("", None, gong_command_set!()); // Should panic here in debug mode!
     }
 
     /* Command names cannot contain the unicode replacement char (`\u{FFFD}`). Support for handling
@@ -362,6 +362,6 @@ mod tests {
     #[test]
     #[cfg_attr(debug_assertions, should_panic)]
     fn create_cmd_with_rep_char() {
-        let _cmd = Command::new("a\u{FFFD}b", None, gong_command_set_fixed!()); // Should panic here in debug mode!
+        let _cmd = Command::new("a\u{FFFD}b", None, gong_command_set!()); // Should panic here in debug mode!
     }
 }