+++ lib/src/macros.rs
@@ -21,15 +21,15 @@
 /// # #[macro_use]
 /// # extern crate gong;
 /// # fn main() {
-/// let _ = gong_option_set_fixed!([ gong_longopt!("foo") ], [ gong_shortopt!('a') ]);
+/// let _ = gong_option_set!([ gong_longopt!("foo") ], [ gong_shortopt!('a') ]);
 /// # }
 /// ```
 #[macro_export]
-macro_rules! gong_option_set_fixed {
+macro_rules! gong_option_set {
     ( $long:tt, $short:tt ) => {
         $crate::options::OptionSet { long: &$long, short: &$short }
     };
-    () => { gong_option_set_fixed!([], []) };
+    () => { gong_option_set!([], []) };
 }
 
 /// Constructs a [`CommandSet`](commands/struct.CommandSet.html)
@@ -42,15 +42,15 @@ macro_rules! gong_option_set_fixed {
 /// # #[macro_use]
 /// # extern crate gong;
 /// # fn main() {
-/// let _ = gong_command_set_fixed!([ gong_command!("foo") ]);
+/// let _ = gong_command_set!([ gong_command!("foo") ]);
 /// # }
 /// ```
 #[macro_export]
-macro_rules! gong_command_set_fixed {
+macro_rules! gong_command_set {
     ( $cmds:tt ) => {
         $crate::commands::CommandSet { commands: &$cmds }
     };
-    () => { gong_command_set_fixed!([]) };
+    () => { gong_command_set!([]) };
 }
 
 /// Constructs a [`LongOption`](options/struct.LongOption.html)
@@ -109,8 +109,8 @@ macro_rules! gong_shortopt {
 /// # #[macro_use]
 /// # extern crate gong;
 /// # fn main() {
-/// let opts = gong_option_set_fixed!();     // An example (empty) option set
-/// let subcmds = gong_command_set_fixed!(); // An example (empty) command set
+/// let opts = gong_option_set!();     // An example (empty) option set
+/// let subcmds = gong_command_set!(); // An example (empty) command set
 ///
 /// let _ = gong_command!("foo");
 /// let _ = gong_command!("foo", @opts &opts);           // With option set
@@ -121,13 +121,13 @@ macro_rules! gong_shortopt {
 #[macro_export]
 macro_rules! gong_command {
     ( $name:expr ) => {
-        $crate::gong_command!($name, @opts $crate::gong_option_set_fixed!(), @cmds $crate::gong_command_set_fixed!())
+        $crate::gong_command!($name, @opts $crate::gong_option_set!(), @cmds $crate::gong_command_set!())
     };
     ( $name:expr, @opts $opts:expr ) => {
-        $crate::gong_command!($name, @opts $opts, @cmds $crate::gong_command_set_fixed!())
+        $crate::gong_command!($name, @opts $opts, @cmds $crate::gong_command_set!())
     };
     ( $name:expr, @cmds $sub_cmds:expr ) => {
-        $crate::gong_command!($name, @opts $crate::gong_option_set_fixed!(), @cmds $sub_cmds)
+        $crate::gong_command!($name, @opts $crate::gong_option_set!(), @cmds $sub_cmds)
     };
     ( $name:expr, @opts $opts:expr, @cmds $sub_cmds:expr ) => {
         $crate::commands::Command {