+++ lib/src/parser.rs
@@ -58,8 +58,8 @@ pub struct Parser<'r, 's: 'r> {
 impl<'r, 's: 'r> Default for Parser<'r, 's> {
     fn default() -> Self {
         Self {
-            options: &gong_option_set_fixed!(),
-            commands: &gong_command_set_fixed!(),
+            options: &gong_option_set!(),
+            commands: &gong_command_set!(),
             settings: Settings::default(),
         }
     }
@@ -120,7 +120,7 @@ impl<'r, 's: 'r> Parser<'r, 's> {
     pub fn new(options: &'r OptionSet<'r, 's>, commands: Option<&'r CommandSet<'r, 's>>) -> Self {
         Self {
             options: options,
-            commands: commands.unwrap_or(&gong_command_set_fixed!()),
+            commands: commands.unwrap_or(&gong_command_set!()),
             settings: Settings::default(),
         }
     }