+++ lib/tests/common/base.rs
@@ -14,7 +14,7 @@ use gong::options::OptionSet;
 use gong::commands::CommandSet;
 
 /// A base set of options for common usage in tests
-static BASE_OPTS: OptionSet = gong_option_set_fixed!(
+static BASE_OPTS: OptionSet = gong_option_set!(
     [
         gong_longopt!("help"),
         gong_longopt!("foo"),
@@ -36,12 +36,12 @@ static BASE_OPTS: OptionSet = gong_option_set_fixed!(
 );
 
 /// A base set of commands for common usage in tests
-static BASE_CMDS: CommandSet = gong_command_set_fixed!([
+static BASE_CMDS: CommandSet = gong_command_set!([
     gong_command!("foo"), // For command/option name clash testing
     gong_command!("add"),
     gong_command!("commit"),
     gong_command!("push",
-        @opts &gong_option_set_fixed!(
+        @opts &gong_option_set!(
             [
                 gong_longopt!("help"),
                 gong_longopt!("tags"),
@@ -50,9 +50,9 @@ static BASE_CMDS: CommandSet = gong_command_set_fixed!([
                 gong_shortopt!('h'),
             ]
         ),
-        @cmds gong_command_set_fixed!([
+        @cmds gong_command_set!([
             gong_command!("origin",
-                @opts &gong_option_set_fixed!(
+                @opts &gong_option_set!(
                     [
                         gong_longopt!("help"),
                         gong_longopt!("force"),
@@ -65,7 +65,7 @@ static BASE_CMDS: CommandSet = gong_command_set_fixed!([
         ])
     ),
     gong_command!("branch",
-        @opts &gong_option_set_fixed!(
+        @opts &gong_option_set!(
             [
                 gong_longopt!("help"),
                 gong_longopt!("sorted"),
@@ -74,11 +74,11 @@ static BASE_CMDS: CommandSet = gong_command_set_fixed!([
                 gong_shortopt!('h'),
             ]
         ),
-        @cmds gong_command_set_fixed!([
+        @cmds gong_command_set!([
             gong_command!("add"),
             gong_command!("del",
-                @opts &gong_option_set_fixed!([], []),
-                @cmds gong_command_set_fixed!([
+                @opts &gong_option_set!([], []),
+                @cmds gong_command_set!([
                     // Note, the names here are chosen to be different to those below for greater
                     // assurance that a match is made from this set, not the sibling below.
                     gong_command!("locally"),
@@ -86,7 +86,7 @@ static BASE_CMDS: CommandSet = gong_command_set_fixed!([
                 ])
             ),
             gong_command!("list",
-                @opts &gong_option_set_fixed!(
+                @opts &gong_option_set!(
                     [
                         gong_longopt!("help"),
                         gong_longopt!("show-current"),
@@ -94,7 +94,7 @@ static BASE_CMDS: CommandSet = gong_command_set_fixed!([
                     ],
                     []
                 ),
-                @cmds gong_command_set_fixed!([
+                @cmds gong_command_set!([
                     gong_command!("local"),
                     gong_command!("remote"),
                 ])